+++ sdk/ml/azure-ai-ml/tests/component/unittests/test_parallel_component_schema.py
@@ -16,6 +16,8 @@ from azure.ai.ml.constants import BASE_PATH_CONTEXT_KEY
 from azure.ai.ml.entities import ParallelComponent
 from azure.ai.ml.entities._assets import Code
 
+from .._util import _COMPONENT_TIMEOUT_SECOND
+
 
 def load_component_entity_from_yaml(
     path: str,
@@ -81,6 +83,7 @@ def load_component_entity_from_rest_json(path) -> ParallelComponent:
     return internal_component
 
 
+@pytest.mark.timeout(_COMPONENT_TIMEOUT_SECOND)
 @pytest.mark.unittest
 class TestParallelComponent:
     def test_serialize_deserialize_basic(self, mock_machinelearning_client: MLClient):
@@ -89,12 +92,6 @@ class TestParallelComponent:
         rest_path = "./tests/test_configs/components/basic_parallel_component_score_rest.json"
         target_entity = load_component_entity_from_rest_json(rest_path)
 
-        # backend add optional=False and port name to inputs/outputs so we add it here manually
-        for name, input in component_entity.inputs.items():
-            input["optional"] = str(input.get("optional", False))
-            input["name"] = name
-        for name, output in component_entity.outputs.items():
-            output["name"] = name
         # skip check code and environment
         component_dict = component_entity._to_dict()
         assert component_dict["id"]