+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.2.0b6 (Unreleased)
 
 ### Features Added
+- Added `TargetAuthorization` of type `dict[str, str]`.
 
 ### Breaking Changes
 - Renamed `begin_create_composed_model()` on `DocumentModelAdministrationClient` to `begin_compose_model()`.
@@ -11,6 +12,8 @@
 - Renamed `DocumentModelInfo` model to `DocumentModelSummary`.
 - Renamed `DocumentModel` to `DocumentModelInfo`.
 - Removed `continuation_token` keyword from `begin_analyze_document()` and `begin_analyze_document_from_url()` on `DocumentAnalysisClient` and from `begin_build_model()`, `begin_compose_model()` and `begin_copy_model_to()` on `DocumentModelAdministrationClient`.
+- Changed return type of `get_copy_authorization()` from `dict[str, str]` to `TargetAuthorization`.
+- Changed expected `target` parameter in `begin_copy_to()` from `dict[str, str]` to `TargetAuthorization`.
 
 ### Bugs Fixed